+++ tests/integration/cases/dict_mixed_int_widths/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,8 @@
+object Fixture_dict_mixed_int_widths_Scala_heterogeneous_strategy_tuple {
+case class Record0(a: Int, b: Long, c: String)
+val my_data = Record0(
+    a = 1,
+    b = 3000000000L,
+    c = "x",
+)
+}