+++ tests/integration/cases/heterogeneous_list_with_string/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,6 @@
+object Fixture_heterogeneous_list_with_string_Scala_heterogeneous_strategy_tuple {
+val my_data = (
+    "hello",
+    42,
+)
+}