+++ tests/integration/cases/nested_mixed_inner/Scala_heterogeneous_strategy_tuple_inner@v3.scala
@@ -1,0 +1,6 @@
+object Fixture_nested_mixed_inner_Scala_heterogeneous_strategy_tuple_inner {
+val my_data = List(
+    List(1, "a"),
+    List(2, "b"),
+)
+}