+++ tests/integration/cases/nested_sequences/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,6 @@
+object Fixture_nested_sequences_Scala_heterogeneous_strategy_tuple {
+val my_data = List[List[List[Int]]](
+    List[List[Int]](List[Int](1, 2), List[Int](3, 4)),
+    List[List[Int]](List[Int](5)),
+)
+}