+++ tests/integration/cases/tuple_record_seq_sibling/Scala_heterogeneous_strategy_tuple@v3.scala
@@ -1,0 +1,16 @@
+object Fixture_tuple_record_seq_sibling_Scala_heterogeneous_strategy_tuple {
+case class Record0(scores: List[Int], args: (Int, String, String, Int))
+val my_data = Record0(
+    scores = List[Int](
+        10,
+        20,
+        30,
+    ),
+    args = (
+        1,
+        "email",
+        "a@gmail.com",
+        100,
+    ),
+)
+}